writecache: use two stage detach
Avoid flushing, and potentially blocking for a long time,
in suspend by using the cleaner setting.

To detach the writecache, first set the cleaner option on the
writecache LV without detaching the writecache.  Then return
to the top level of the command, releasing the VG and the VG lock.
From there, periodically check the progress of the cleaner by
locking/reading the VG and checking kernel status.  Once the
cleaner has finished flushing, detach the writecache from the LV.